+++ csharp/cwrapper/cwrapper.c
@@ -4,7 +4,7 @@
 #include <sys/io.h>
 
 
-bool port_out_32(uint32_t port, uint64_t value)
+bool port_out_32(uint16_t port, uint32_t value)
 {
 	if (ioperm(port, 4, 1) < 0 || ioperm(0x80, 1, 1) < 0) {
 		return false;
@@ -14,7 +14,7 @@ bool port_out_32(uint32_t port, uint64_t value)
 	return true;
 }
 
-uint32_t port_in_32(uint32_t port)
+uint32_t port_in_32(uint16_t port)
 {
 	if (ioperm(port, 4, 1) < 0) {
 		return (uint32_t) -1; // same as if absent
